Too much for one commit...
I really shouldn't be commiting all this at once but I'm about to
redesign my website's code again. So, I think it's ok this once.

Thing's I was working on before running into issues:
	- Adding custom reactivity CSS for skins
	- Finishing skins
	- Added a few "souls" for Ashen
	- Added meta.SVG for help with metadata
	- Trying to imporove code readability...

So, I haven't had to rewrite something in a while and this is going to be an undertaking...
For the most part, the issue right now is stemming from muscles.js, the code that I've written
in isn't well managed and isn't very portable. With that said, I know not all code needs to be
portable or managed within the confines of OOP. But I want the code to be as readable as possible
for both my sake and the sake of anybody who reads my code. It's just anoying to read jank code or
sloppy code and It's a hastl I know all to much about. Besides, If it's readable, then maybe
someone will learn something from me right?